flask-vue/requirements.txt
bcrypt==3.1.7
blinker==1.4
cffi==1.14.0
click==7.1.2
dnspython==1.16.0
email-validator==1.1.1
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.4.3
Flask-WTF==0.14.3
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pdfkit==0.6.1
# Pillow==7.2.0
Pillow
pycparser==2.20
python-dotenv==0.14.0
six==1.15.0
SQLAlchemy==1.3.18
Werkzeug==1.0.1
WTForms==2.3.1
# bleach==5.0.0
bleach
# Markdown==3.4.1
Markdown
# flask-moment==1.0.5
flask-moment
bootstrap-flask
pandas